Restructured module layout.
