+++ quickterm.cabal
@@ -57,7 +57,7 @@ source-repository head
 
 library
   -- Modules exported by the library.
-  exposed-modules: Quickterm
+  exposed-modules: Quickterm, Help, Deserializer, Demo, CanMarshall
 
   -- Modules included in this library but not exported.
   -- other-modules:
